dmesg: Add --noctrl to disable printing control characters from kmsg
This is similar to --noescape, but only disables printing control
characters, as they may distort the dmesg output. For example, this line
in /dev/kmsg...

    6,2055,346524498636,-;usb 1-2.4: SerialNumber: \x08\x08\x08\x08\x04\x04\x04\x04

...shows like this in dmesg:

    usb 1-2.4: SerialNumb^D^D^D^D

It's nice to have "special" characters included in dmesg (for example, I
run Chinese-language applications which may end up with lines needing
hex unmangling), but backspaces and EOTs are much less likely to be
useful.

It's possible that this could just go with the above isprint/isspace
checks, but I'm not sure if anyone already relies on this behaviour, so
I've put it behind a flag for now.

Signed-off-by: Chris Down <chris@chrisdown.name>